+++ CHANGELOG.md
@@ -1,11 +1,23 @@
 # CHANGELOG
 
 ## 17.11.2017
+## 20.11.2017
 ### Added
 - Tiled collision support 
 
 ### Changed
   - Improved collision performance
+
+- struct CollisionData
+    - Includes pointer to the other Collidable
+    - Includes the closest normal of the owner
+        - Center-to-center -vector if owner is a circle
+- GetNormals() for Collidable
+
+### Changed
+- Collidable::GetCollisions() now returns a vector of CollisionData
+
+
 
 ## 11.11.2017
 ### Added